patent-agent/requirements.txt
#requirements.txt
groq
chromadb
python-dotenv
